Home page (#6)
* created link between dashboard and home website

* improvements to the header

* User is not tracked with the token in local storage

+++ src/app/app.component.ts
@@ -2,6 +2,9 @@ import {Component} from '@angular/core';
 import {HttpService} from "../services/http.service";
 import {Router, Event, NavigationEnd} from '@angular/router';
 import {MatSnackBar} from "@angular/material/snack-bar";
+import * as http from "http";
+import jwtDecode from "jwt-decode";
+import {User} from "./types/user";
 
 
 @Component({
@@ -22,23 +25,35 @@ export class AppComponent {
         this.currentRoute = event.url;
       }
     });
+
+    let token = localStorage.getItem('token') as any;
+    if (token){
+      let decoded = jwtDecode(token) as any;
+      let currentDate = new Date();
+      if (decoded.exp){
+        let expiry =  new Date(decoded.exp*1000);
+
+        if (currentDate<expiry){
+          this.http.ReadUserFromStorage(token);
+        } else {
+          localStorage.clear();
+          alert("You have been logged, login to continue.")
+          //TODO
+        }
+      }
+    }
   }
 
-  route() {
-    this.router.navigate(['./Home'])
-  }
+
 
   logOut() {
     this.router.navigate(['Home']).then(() => {
       this.snackBar.open('You have now been logged out', undefined, {duration: 3000})
-      localStorage.clear();
     })
-    this.http.user.name = '';
     localStorage.clear();
+    this.http.IsUser = false;
+    this.http.user.UserName  = '';
+    this.http.user.Email = '';
+    this.http.user.Id = '';
   }
-
-  logIn() {
-    this.router.navigate(['Login']);
-  }
-
 }
